fsck.gfs2: Make pass1 undo its work for unrecoverable inodes
When pass1 discovers a dinode that has so much damage it finally
gives up on it, it was invalidating all the inode's metadata blocks.
The problem with with doing that is that it often gave up on the
dinode half-way through processing its metadata, so some of the
metadata blocks were marked as some block type n the blockmap (such
as indirect blocks or duplicate reference) while the unprocessed
blocks were not marked at all, leaving them unaccounted for.
Later, marking them as invalid caused subsequent processing of
other dinodes that referenced those blocks to think they were
duplicate references, often with disasterous consequences.  Often
the duplicate processing code in pass1b would favor the first
("invalidated") reference and delete the second ("good") reference,
which means it introduced corruption.  The proper thing to do is
to undo its blockmap designations, and stop at the same place it
did when it gave up on the dinode.  This patch introduces a new
set of "undo" metadata processing functions to accomplish this task.
These undo_check_metalist and related functions MUST be kept in sync
with pass1's normal pass1_fxns functions.  In other words, they must
stop at the same block of metadata and return the same error to
ensure anything that pass1_fxns does is properly undone by the
corresponding undo functions.

rhbz#455300