fsck.gfs2: Overhaul duplicate reference processing
This patch is a major overhaul of the duplicate reference processing
that had so many problems.  Before this patch, pass1 would flag a
block as a duplicate reference, then pass1b would go to great lengths
to find all dinodes that reference the block in question.  Then it
tried to figure out which reference to keep and which to discard.
This was slow and prone to errors.  With this patch, when pass1
discovers a block that's already got a blocktype, it keeps tabs on
that block and what dinodes reference it from then on.  That means
pass1b only needs to find one reference: The original reference that
caused it to become not free.  Once that is found, it can stop
looking.  Furthermore, this patch introduces two linked lists of
inode references: valid references and invalid references.  When
an inode is determined to be bad or invalid, its duplicate block
references are moved from the valid to the invalid reference list.
That way, pass1b can favor the most valid references over the invalid
references, and it's more likely to keep good data and discard trash.

rhbz#455300