fsck.gfs2 fails on root fs: Device X is busy.
In a previous commit, we added an O_EXCL when opening the
device.  This was done to prevent other nodes from interacting
with the file system while it was being checked.  However, this
turns out to be a problem when the root file system is GFS2.
In that case, it's proper to allow the fsck to complete normally
and if errors are found, tell vfs to flush its cache after
corrections are made.

rhbz#557128