GFS2: fsck.gfs2 segfault - osi_tree "each_safe" patch
There are a few places in fsck.gfs2 where each element of the
new rbtrees are processed.  Some of those places delete
items from the rbtree, which means I need to use the equivalent
of osi_list_foreach_safe.  In other words, I needed to prevent
rbtree deletes from interfering with the next() function.
The result was a segfault in fsck.gfs2 in a few places.  This
patch implements the "safe" rbtree traversal and fixes the problem.

rhbz#574215