cman: Add improved cluster_id hash function
Add a new and vastly improved function for converting the cluster name
into a 16bit ID. This is currently disabled by default but can be enabled
with the stanza

<cman hash_cluster_id="yes"/>

in cluster.conf

We might make this the default in future .. watch this space. Or some space
closely related to this one that's less intimidating to read on a Friday
afternoon before you've had tea.

Thanks to Steven Dake for finding and slightly modifying the hash function.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>