rgmanager: Kill processes correctly w/ force_unmount
The killMountProcesses function was written about 10 years ago.
It was designed to work with lsof or fuser, and to log messages
for each process killed.  This is not a bad idea.  The problem
is that parsing the output of either is and error-prone,
particularly when mountpoints are similar to other directories
on the system.

A far less error-prone method to cleaning up a mount point is to
use 'fuser -kvm' on it.  Not only is this less error-prone, it's
a good bit faster at doing its job than iterating through output
in a shell script.

This patch makes force_unmount very reliable at killing the correct
processes, but we lose the logging functionality.  It is a fair
trade-off because there have been several bugs in the
killMountProcesses function over the years which have caused several
problems.

Resolves: bz555901 bz582754

Signed-off-by: Lon Hohberger <lhh@redhat.com>