qdiskd: Change votes instead of cman-visible status
With master-wins mode, we were advertising availability
of votes to cman.  This caused cman_tool nodes output to
claim the quorum disk was offline.  Since the initscript
expects the quorum disk to be reported as 'online', this
caused strange behaviors.

This fix makes us always advertise our votes as available.
Instead of presenting the votes vs. not presenting, we
simply present 0 votes when not master and 1 vote if
master.

Resolves: bz#576871

Signed-off-by: Lon Hohberger <lhh@redhat.com>