halvm: Fix bug 506587: lvm agent incorrectly reports vg is in volume_list
If the name of the VG controlled by rgmanager is a substring of the name
of the node, the HA-LVM script will complain of an improper setup.

The fix is to also look for the associated quotes that are necessary when
specifying strings in lvm.conf.  Thanks to John Ruemker for the patch.