cman: fix quorum recalculation when a node is externally killed
This patch fixes a condition where quorum would be incorrectly recalculated
if a node was removed from the cluster using the cman_kill_node call.
This call is used by qdisk and can cause a cluster to continue operating
when too few nodes are left in the cluster for a legitimate quorum.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>