resource-agents: Make vm.sh use stop/start timeouts
Startup timeouts are handled by calling the status_program
(if specified) every few seconds until either the timeout is
reached or the status program returns a successful result.

Because the combined VM boot + application time was the
basis for the 5 minute check interval for depth=10 (where
the status_program is called), it has been reduced to 1
minute intervals.

The startup timeout, since it may now be waiting for
services within the VM, has been increased to 5 minutes
from 20 seconds to match the previous status check timing
functionality.

This patch fixes the previous patch's inadequate
handling of start timeouts.

Resolves: bz606754

Signed-off-by: Lon Hohberger <lhh@redhat.com>