cman: Deprecate specifying config version to cman_tool
Historically, updating the configuration file required
two discrete steps:

  1) send configuration to all nodes, and

  2) tell CMAN about the new configuration version

There has always been a possibility that an incorrect
version number could be sent to CMAN, causing it to have
a different version than the one in cluster.conf.

In the STABLE3 and later branches as of commit 75fb0324,
the necessity for specifying the version # has been
removed; users could instead just use '-r 0' to use the
configuration version in the current cluster.conf.

This patch deprecates specification of the configuration
version number from the cman_tool command line while
attempting to retain command-line tool compatibility.
That is, you may specify '-r123', but it will be ignored
and a warning will be printed.

Resolves: rhbz#617163

Signed-off-by: Lon Hohberger <lhh@redhat.com>