config: more cman_tool config reload cleanup
when issuing cman_tool version on the local node, and the local node
cannot read the new config, there is very little point to keep trying
loading it (since we are the issuing node).

make sure that if the local node cannot read the config, we stop immediately
and report a proper error down to cman_tool to display.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>