dlm_controld: fix plock owner in checkpoints
The wrong plock resource owner is written into checkpoints
when plock_ownership is 0.  This causes a node that mounts
the fs to have incorrect owner values, which cause the
plock operations to permanently hang.

This bug seems to have existed since the plock code was originally
copied into dlm_controld from gfs_controld.  As part of the copy,
there were some small code changes.  One was to always include the
resource owner in the checkpoint data, instead of only including it
when plock_ownership was 1.  The owner was then written and read
incorrectly when plock_ownership was 0.

bz 618814

Signed-off-by: David Teigland <teigland@redhat.com>