cman: Make qdiskd exit if removed from configuration
If qdiskd was previously configured and a configuration
update removes the device & label fields, qdiskd will
now set its votes to 0, unregister from CMAN, and exit.

Correct logging about device registration and
configuration errors is also required in order for
users to understand what is going on when this happens
since the current logging from CMAN is a bit unclear.

Resolves: rhbz#620679

Signed-off-by: Lon Hohberger <lhh@redhat.com>