dlm_controld: fix plock signature in stored message
When confchg's change the ckpt_node between plock checkpoints
being written, and the plocks_stored message being received,
the plocks_stored message is resent by the new ckpt_node that
didn't write the checkpoint.

The new ckpt_node needs to save the plock checkpoint signature
that was sent in the first plocks_stored message from the
checkpoint writer, and reuse it when sending the new
plocks_stored method.  Otherwise the plock signature in the
stored message will be invalid.

bz 623816

Signed-off-by: David Teigland <teigland@redhat.com>