dlm_controld: fix plock owner syncing
- The R_GOT_UNOWN flag was not always being set on resources when
the owner was set to 0.  This would cause subsequent syncing of
plock state to write the incorrect owner into the checkpoint.

- Plocks were being written into the checkpoint unnecessarily for
owned resources in cases where the owner written in the checkpoint
was not the same as r->owner.

- Set a few other flags in r->flags to help debug future problems.

- Do more error checking and verifying of checkpointed state when
checkpoints are being written and read.

- Handle errors during plock syncing by disabling plock operations,
instead of ignoring or running with incorrect state.

bz 617306

Signed-off-by: David Teigland <teigland@redhat.com>