dlm_controld: ignore plocks until checkpoint time
After our join confchg, we need to ignore plock messages
until the point in time where the ckpt_node saves plock
state (final start message received).  At that time we need
to shift from ignoring plock messages to saving plock
messages that will be applied on top of the plock state that
is read from the checkpoint.  The code was not ignoring
plock messages during the first stage of the process, which
led to incorrect plock state.

bz 623810

Signed-off-by: David Teigland <teigland@redhat.com>