Updating /proc/mounts and /etc/mtab with mount args for GFS2 fs
This patch changes the way the gfs2 mount helper, mount.gfs2,
updates the /etc/mtab file.  Before, it was copying the line
from /proc/mounts directly into /etc/mtab after a successful
mount.  The problem is, that may be mapped to a different device
by the device-mapper layer, which meant it did not match the
device properly if a subsequent mount was done for the same device.
This patch uses the device specified on the mount.gfs2 command
line argument (which normally comes from the mount command), already
translated properly as it expects for subsequent mount attempts.

rhbz#632595